wrapping board in its context
