+++ pinned-react/src/context/BoardContext.js
@@ -1,9 +1,11 @@
-import React, { createContext, useState } from 'react'
+"use client"
+
+import React, { createContext, useContext, useState } from 'react'
 
 // context that will be use throughout the app
 const BoardContext = createContext(null);
 
-const BoardContextProvider = ({ children }) => {
+export const BoardContextProvider = ({ children }) => {
     const [board, setBoard] = useState(null); 
 
   return (
@@ -14,4 +16,13 @@ const BoardContextProvider = ({ children }) => {
   )
 }
 
-export {BoardContext, BoardContextProvider};
+
+export function useBoardContext() {
+  const context = useContext(BoardContext)
+
+  if (!context) {
+    throw new Error("Board context must be used within BoardContextProvider");
+  }
+
+  return context;
+}
