+++ pinned-react/src/pages/Board/EventsPage.js
@@ -2,12 +2,13 @@ import { useState, React, useEffect } from 'react'
 import NewEventSystem from '../../components/board/new event/NewEventSystem'
 import axios from 'axios';
 import { useParams } from 'react-router-dom';
+import { useBoardContext } from '../../context/BoardContext';
 
 
 const EventsPage = () => {
   const { id } = useParams()
-
   const [events, setEvents] = useState(null);
+  const { board, setBoard } = useBoardContext();
 
   useEffect(() => {
     axios.get(`http://localhost:8080/api/events/of-board/${id}`)
@@ -21,33 +22,34 @@ const EventsPage = () => {
    }, [id])
 
   return (
-    <div className="h-screen w-screen bg-gray-50 text-gray-950 dark:[#282c34]">
-            <h1 className='text-center border w-full border-black bg-actionOrange h-[3rem]'>
-                Temp Heading
-            </h1>
-
-          <div className="flex-corl">
-            {events && <div className="flex gap-3 mt-5 ml-5">
-            {events.map((event, index) => (
-           
-                <div className="w-[24rem] h-[24rem] border border-actionOrange" key={index}>
-                    <h1 className="mb-2">{event.title}</h1>
-                    <p className="mb-2">{event.description}</p>
-                    <ul className="mb-2">
-                      {event.tags.map((tag, i) => (
-                          <li key={i}>{tag}</li>
-                      ))}
-                    </ul>
-                    <p className="mb-2">{event.createdAt}</p>
-                    <p className="mb-2">{event.updatedAt}</p>
-                </div>   
-            ))}
-         </div>}
-        </div>
-
-
+    <div>
+      {board ? (
+        <div className="h-screen w-screen bg-gray-50 text-gray-950 dark:[#282c34]">
+              <div className="flex-corl">
+                {events && <div className="flex gap-3 mt-5 ml-5">
+                {events.map((event, index) => (
+              
+                    <div className="w-[24rem] h-[24rem] border border-actionOrange" key={index}>
+                        <h1 className="mb-2">{event.title}</h1>
+                        <p className="mb-2">{event.description}</p>
+                        <ul className="mb-2">
+                          {event.tags.map((tag, i) => (
+                              <li key={i}>{tag}</li>
+                          ))}
+                        </ul>
+                        <p className="mb-2">{event.createdAt}</p>
+                        <p className="mb-2">{event.updatedAt}</p>
+                    </div>   
+                ))}
+            </div>}
+            </div>
             <NewEventSystem />
-        
+        </div>
+      ) : (
+        <div>
+            <h1>Board not found.</h1>
+          </div>
+      )}
     </div>
   )
 }
