--- pinned-react/src/pages/CalendarPage.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const CalendarPage = () => {
-  return (
-    <div>
-      <h1>You are on the Calendar Page</h1>
-    </div>
-  )
-}
-
-export default CalendarPage;
